Remove unnecessary data fetching to improve loading times
